vyatta-openvpn: persistent tunnels in server mode and allow full reset
Allow OpenVPN tunnels (vtun) that are configured for server mode
operation to be set as persistent (TUN/TAP interfaces aren't closed
and reopened on a reset).  Check if the tunnel is configured for
persistence when 'reset openvpn interface' is run from the CLI, if it
then prompt the user if an interface reset is required along with the
tunnel reset (defaults to yes).

Bug #7 http://bugzilla.vyos.net/show_bug.cgi?id=7
Bug #313 http://bugzilla.vyos.net/show_bug.cgi?id=313