T1388: Fix client connections with password and cert authentication
If an OpenVPN client-mode connection is setup with password authentication,
the certificate will not be passed to the server, which prevents the tunnel
from being brought up. A possible workaround is to set the certificate and key
manually through the "openvpn-option" setting.

(cherry picked from commit cdd5d072f5a722156a06e600b5742e8bc3854b55)